improved user experience

+++ download_file_handler.py
@@ -23,39 +23,44 @@ class DownloadFileHandler(FileSystemEventHandler):
         src_path = event.src_path
         if (any(name in src_path for name in temp_file_names) or src_path in self.file_path or src_path in self.file_cache):
             # ignore temporary files
-            print("ignored")
             return
-        print(f"new file -  {src_path} created!")
+        print(f"new file -  {src_path} downloaded!")
         file_name = self.get_filename(src_path)
         similar_files = self.get_similar_files(file_name)
-        print(f"similar files: {similar_files}")
-        keyword = self.get_keyword(similar_files)
-        dest_folder = self.file_path + keyword
-        print(f"file name: {file_name}")
-        is_folder_created = self.is_folder_created(file_name)
-        try:
-            if (is_folder_created == ""): #folder has not been created
-                self.create_folder(keyword)
-                self.move_files(similar_files, dest_folder)
-                #write new folder name to new_folders.txt
-                self.write_folder(keyword)
-            else: #folder is already created, move file to this folder
-                self.move_files([src_path], dest_folder)
-        except shutil.Error as e:
-            duplicate_file = self.extract_duplicate_file(e.args[0], dest_folder)
-            print(f"duplicate filename: {duplicate_file}")
-            if (filecmp.cmp(self.file_path + duplicate_file, dest_folder + "/" + duplicate_file)):
-                inp = input(f"duplicate files {duplicate_file} found in Downloads and {dest_folder}, delete the one in Downloads? [y]/[n]")
-                while (inp not in ["y", "n"]):
-                    print("please input one of [y] or [n]")
-                if (inp == 'y'):
-                    self.remove_file(self.file_path + duplicate_file)
-        self.file_cache.append(src_path)
+        print(f"length of similar files: {len(similar_files)}")
+        if (len(similar_files) > 1): # if there are multiple similar files that need to be organized
+            print(f"similar files: {similar_files} found in Downloads")
+            keyword = self.get_keyword(similar_files)
+            # print(f"keyword: {keyword}")
+            dest_folder = self.file_path + keyword
+            is_folder_created = self.is_folder_created(file_name)
+            # print(f"is folder created : {is_folder_created}")
+            try:
+                if (is_folder_created == ""): #folder has not been created
+                    self.create_folder(keyword)
+                    self.move_files(similar_files, dest_folder)
+                    #write new folder name to new_folders.txt
+                    self.write_folder(keyword)
+                else: #folder is already created, move file to this folder
+                    # print(f"dest_folder: {dest_folder}")
+                    self.move_files([src_path], dest_folder)
+            except shutil.Error as e:
+                duplicate_file = self.extract_duplicate_file(e.args[0], dest_folder)
+                # print(f"duplicate file: {duplicate_file}")
+                if (filecmp.cmp(self.file_path + duplicate_file, dest_folder + "/" + duplicate_file)):
+                    # print("line 46")
+                    inp = input(f"duplicate files {duplicate_file} found in Downloads and {dest_folder}, delete the one in Downloads? [y]/[n]\n")
+                    while (inp not in ["y", "n"]):
+                        print("please input one of [y] or [n]")
+                        inp = input()
+                    if (inp == 'y'):
+                        self.remove_file(self.file_path + duplicate_file)
+                        print(f"{self.file_path + duplicate_file} deleted")
+            self.file_cache.append(src_path)
         return
     
     def extract_duplicate_file(self, err_msg, dest_folder):
         duplicate_filename = re.sub(rf"^Destination path '{dest_folder}/|' already exists$", "", err_msg)
-        # duplicate_filename = re.match()
         return duplicate_filename
 
     def remove_file(self, file_name):
@@ -98,10 +103,10 @@ class DownloadFileHandler(FileSystemEventHandler):
     # average fuzz similarity score between new file and each existing file in the directory 
     # if avg score > threshold -> belongs to that folder 
     # if there is no such folder return an empty string
-    #TODO return the HIGHEST folder similarity score
+    # this function lazily returns the first similar folder it finds, not necessarily the MOST similar
     def is_folder_created(self, new_file):
         new_folders = self.read_folders()
-        print(f"new_folders: {new_folders}")
+        # print(f"new_folders: {new_folders}")
         for folder in new_folders:
             similarity_score = 0
             folder_size = 0
@@ -123,7 +128,7 @@ class DownloadFileHandler(FileSystemEventHandler):
         for entry in entries:
             if (self.check_similarity(new_file, entry.name)):
                 similar_files.append(entry)
-        print(f"similar_files: {similar_files}")
+        # print(f"similar_files: {similar_files}")
         return similar_files
 
     
@@ -136,16 +141,18 @@ class DownloadFileHandler(FileSystemEventHandler):
     # extract file name from the full path in src_path
     def get_filename(self, src_path):
         result = src_path.split(self.file_path)[1]
-        print(f"file name: {result}")
+        # print(f"file name: {result}")
         return result
     
     # extract keyword from a list of similar file names
     # finds the most common substring and returns it
     # remove non-alphanumeric characters at beginning or end
     def get_keyword(self, similar_files):
+        #need to convert all file names to lower case for accurate keyword
+        similar_files = [s.name.lower() for s in similar_files]
         n = len(similar_files)
         # use first filename as reference
-        s = similar_files[0].name
+        s = similar_files[0]
         l = len(s)
         result = ""
         for i in range(l): 
@@ -155,7 +162,7 @@ class DownloadFileHandler(FileSystemEventHandler):
                 common_files = 1
                 for k in range(1, n):
                     # Check if the substring is common to filenames
-                    if substr not in similar_files[k].name:
+                    if substr not in similar_files[k]:
                         break
                     else:
                         common_files += 1
